Thappar Hackathon/voyageai-app-main/voyageai-app-main/server/requirements.txt
django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
mysqlclient==2.2.4
google-generativeai==0.6.0
requests==2.32.3
pymysql==1.1.1

